+++ hypha/apply/projects/models/__init__.py
@@ -13,7 +13,6 @@ from .project import (
     ProjectSettings,
     ProjectSOWForm,
 )
-from .report import Report, ReportConfig, ReportPrivateFiles, ReportVersion
 
 __all__ = [
     "Project",
@@ -28,10 +27,6 @@ __all__ = [
     "ContractPacketFile",
     "DocumentCategory",
     "ContractDocumentCategory",
-    "Report",
-    "ReportVersion",
-    "ReportPrivateFiles",
-    "ReportConfig",
     "Invoice",
     "SupportingDocument",
 ]
