+++ hypha/apply/projects/reports/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ReportsConfig(AppConfig):
+    name = "hypha.apply.projects.reports"
+    label = "project_reports"
