+++ hypha/apply/projects/reports/templates/reports/report_detail.html
