+++ hypha/apply/projects/templates/application_projects/project_detail.html
@@ -82,7 +82,7 @@
                 {% project_can_have_report object as can_have_reports %}
                 {% if can_have_reports %}
                     <div class="wrapper wrapper--outer-space-large">
-                        {% include "application_projects/includes/reports.html" %}
+                        {% include "reports/includes/reports.html" %}
                     </div>
                 {% endif %}
