+++ hypha/apply/projects/utils.py
@@ -1,3 +1,6 @@
+import textwrap
+
+from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 from django_file_form.uploaded_file import PlaceholderUploadedFile
 
@@ -31,6 +34,17 @@ from .models.project import (
     PROJECT_STATUS_CHOICES,
     PAFReviewersRole,
 )
+
+PROJECT_TITLE_TEMPLATE = "{text} <span class='text-fg-muted'>#{application_id}</span>"
+
+
+def get_project_title(project):
+    text = textwrap.shorten(project.title, width=40, placeholder="…")
+    return format_html(
+        PROJECT_TITLE_TEMPLATE,
+        text=text,
+        application_id=project.application_id,
+    )
 
 
 def no_pafreviewer_role():
